requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
et-xmlfile==1.1.0
filelock==3.15.1
flatbuffers==24.3.25
fpdf==1.7.2
fsspec==2024.6.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
openpyxl==3.1.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
panda==0.3.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
protobuf==4.25.3
pygame==2.5.2
Pygments==2.18.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rich==13.7.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
tbb==2021.12.0
tenacity==8.4.1
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.6.0
wrapt==1.16.0
XlsxWriter==3.2.0
qrcode==7.3.1